Parsing to LibGDX (2/4)
Now we have more cool stuff which was been parsed. Now we have:

A camera which works in any machine with 800x600 resolution

Collision method with tiles

We also have a restructured way to init levels. In this way we will not have compatibility problems with the machine were we are executing our game

Also have a Color class to facilitate the process of level initiation (once we have a color place system based)

And I think that is it, sorry for this long commits, but since we are parsing the project, I think that that will not be a huge problem. But I promise that, after parsing the game to a game LibGDX system I will do very smaller commits